requirements.txt.txt
django
social-auth-app-django
django-allauth
random
